+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- `CatalogBuilder::add_builder()` method will automatically register interfaces (`#[interface(dyn Iface)]`) for `#[component]`
+
 ## [0.12.4] - 2025-03-26
 ### Fixed
-- Supress `clippy::too_many_arguments` warning in generated functions.
+- Suppress `clippy::too_many_arguments` warning in generated functions.
 
 ## [0.12.3] - 2025-03-26
 ### Changed